Add new `abi-wait-complete` target
# Background

The `start-iso-abi` target starts the ABI installation, but it does not
wait for the installation to complete. Currently the `ib-orchestrate-vm`
repo that uses this repo as a submodule implements its own waiting for
installation to complete logic, which includes running `oc` commands and
looking at the `clusterversion` resource.

# Issue / Solution

This repo should provide a target that waits for the installation to
complete, and it should do it correctly. The openshift-installer
`wait-for` commands do more than just wait, they also populate the admin
`kubeconfig` that is generated with additional ingress certificates that
are created in late stages of cluster finalization. Without them,
running conformance tests using said `kubeconfig` will fail as some
tests assume the presence of these ingress certificates.

# Implementation

Add a new `abi-wait-complete` target that uses the openshift-installer
`wait-for` command to wait for the installation to complete. This target
should be run after the `start-iso-abi` target, and the
`ib-orchestrate-vm` repo will be updated to use this new target instead
of (or on top of) its own logic.

Note that currently, to overcome the AGENT-916 issue, we also run the
`wait-for install-complete` command without the `agent` subcommand. This
is a temporary workaround until AGENT-916 is fixed in all versions we
use.